suno_mastering_agent/requirements.txt
# Core dependencies
playwright>=1.40.0
pydantic>=2.0.0
rich>=13.0.0
click>=8.0.0
python-dotenv>=1.0.0

# AI agent dependencies
browser-use>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-deepseek>=0.1.0
langchain-ollama>=0.2.0
gradio>=5.0.0
pyyaml>=6.0.0

# Optional: additional LLM providers (install as needed)
# langchain-openai>=0.3.0
# langchain-anthropic>=0.3.0
